requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.1.7
django-debug-toolbar==1.11
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
oauthlib==3.0.1
parso==0.3.4
pickleshare==0.7.5
Pillow==5.4.1
pip-chill==0.1.8
prompt-toolkit==2.0.9
Pygments==2.3.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
